Overall model function, fast implementation

Overall model function, fast implementation

antibody_model(
  theta,
  unique_theta_indices,
  unique_biomarker_groups,
  infection_history_mat,
  infection_history_mat_indices,
  indiv_theta_groups,
  possible_exposure_times,
  possible_exposure_times_indices,
  exposure_groups,
  unique_exposure_groups,
  sample_times,
  type_data_start,
  biomarker_groups,
  sample_data_start,
  antibody_data_start,
  nrows_per_sample,
  biomarker_id_indices,
  start_level_indices,
  starting_antibody_levels,
  births,
  antigenic_map_long,
  antigenic_map_short,
  antigenic_distances,
  timevarying_groups = FALSE,
  exponential_waning = FALSE,
  variant_specific_pars = FALSE,
  boost_before_infection = FALSE
)

Arguments

theta

NumericVector, the named vector of model parameters

unique_theta_indices

IntegerVector, the indices of the unique model parameters in the parameter matrix

unique_biomarker_groups

IntegerVector, the biomarker groups represented in the parameter matrix

infection_history_mat

IntegerMatrix, the matrix of 1s and 0s showing presence/absence of infection for each possible time for each individual.

infection_history_mat_indices

IntegerVector, the indices linking infection-history entries to the corresponding model inputs

indiv_theta_groups

IntegerVector, the parameter-group index for each individual

possible_exposure_times

NumericVector, the time periods that the infection history vector corresponds to

possible_exposure_times_indices

IntegerVector, which entry in the melted antigenic map that each possible infection time corresponds to

exposure_groups

IntegerVector, the exposure-group index associated with each possible infection time

unique_exposure_groups

IntegerVector, the exposure groups represented in the model

sample_times

NumericVector, the times that each blood sample was taken

type_data_start

IntegerVector, one entry for each unique individual. Each entry gives the starting index for each individual of the data frame `unique(antibody_data[,c("individual","biomarker_group")])`.

biomarker_groups

IntegerVector, result of `unique(antibody_data[,c("individual","biomarker_group")])$biomarker_group`

sample_data_start

IntegerVector, one entry for each unique individual and observation type combination. Each entry dictates how many indices through sample_times to iterate per individual and observation type (ie. how many sample times does each individual have?)

antibody_data_start

IntegerVector, How many cumulative rows in the antibody data correspond to each unique individual and observation type combination?

nrows_per_sample

IntegerVector, one entry per sample taken. Dictates how many entries to iterate through cum_nrows_per_individual_in_data for each sampling time considered

biomarker_id_indices

IntegerVector, the indices of all measured biomarkers in the melted antigenic map, with one entry per measured biomarker

start_level_indices

IntegerVector, the indices linking each individual, biomarker group, and biomarker ID to its starting antibody level

starting_antibody_levels

NumericVector, the starting antibody levels used for the individual, biomarker group, and biomarker ID combinations

births

NumericVector, the birth times for each individual

antigenic_map_long

arma::mat, the collapsed cross reactivity map for long term boosting, after multiplying by cr_long see create_cross_reactivity_vector

antigenic_map_short

arma::mat, the collapsed cross reactivity map for short term boosting, after multiplying by cr_short, see create_cross_reactivity_vector

antigenic_distances

NumericVector, the collapsed cross reactivity map giving euclidean antigenic distances, see create_cross_reactivity_vector

timevarying_groups

logical, whether demographic groups can vary over time

exponential_waning

logical, whether to use the exponential rather than linear waning and cross-reactivity form

variant_specific_pars

logical, whether parameter blocks are selected by exposure group

boost_before_infection

bool to indicate if calculated antibody level for that time should be before the infection has occurred, used to calculate antibody-mediated immunity

Value

NumericVector of predicted antibody levels for each entry in biomarker_id_indices