Get posterior information infection histories

Finds the median, mean and 95

calculate_infection_history_statistics(
  inf_chain,
  burnin = 0,
  years = NULL,
  n_alive = NULL,
  known_ar = NULL,
  group_ids = NULL,
  known_infection_history = NULL,
  solve_cumulative = FALSE
)

Arguments

inf_chain
the data table with infection history samples from run_MCMC

burnin
if not already discarded, discard burn in from chain (takes rows where sampno > burnin)

years
vector of the epochs of potential circulation

known_ar
data frame of known attack rates, if known.

known_infection_history
data frame of known infection histories.

solve_cumulative
if TRUE, finds the cumulative infection histories for each individual. This takes a while, so is left FALSE by default.

Value

a list of data frames with summary statistics

See also

Examples

data(example_inf_chain)
data(example_antigenic_map)
data(example_titre_dat)
data(example_inf_hist)
strain_isolation_times <- example_antigenic_map$inf_times
## Find number alive in each time period
n_alive <- get_n_alive(example_titre_dat, strain_isolation_times)

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

## Get actual number of infections per time
n_infs <- colSums(example_inf_hist)
## Create data frame of true ARs
known_ar <- n_infs/n_alive
known_ar <- data.frame("j"=strain_isolation_times,"AR"=known_ar,"group"=1)

#> Error in data.frame(j = strain_isolation_times, AR = known_ar, group = 1): arguments imply differing number of rows: 0, 48, 1

## Get true infection histories
known_inf_hist <- data.frame(example_inf_hist)
colnames(known_inf_hist) <- strain_isolation_times

## Need to get group specific n_alive and adjust to correct time frame 
n_alive_group <- get_n_alive_group(example_titre_dat, strain_isolation_times,melt_dat = TRUE)

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Error in melt(n_alive, id.vars = c("group")): could not find function "melt"

n_alive_group$j <- strain_isolation_times[n_alive_group$j]

#> Error in eval(expr, envir, enclos): object 'n_alive_group' not found

results <- calculate_infection_history_statistics(example_inf_chain, 0, strain_isolation_times,
                                                  n_alive=n_alive_group, known_ar=known_ar,
                                                  known_infection_history=known_inf_hist)

#> Padding inf chain...

#> Error in fsetdiff(expanded_values, inf_chain[, c("i", "j", "sampno", "chain_no",     "group")]): could not find function "fsetdiff"