Example parameter table

Example of the parameter table used through serosolver. This data frame is used to control everything related to the model parameters, including values, names, fixed/estimates, and uniform prior bounds.

data(example_par_tab)

Format

A data frame with 17 rows and 10 variables:

names

string names of the model parameters

values

numeric values of the parameters

fixed

binary values indicating if the parameter should be fixed (1) or estimated (0) during the MCMC procedure

lower_bound

lower numeric bound for the parameter during fitting (lower uniform prior bound)

upper_bound

upper numeric bound for the parameter during fitting (upper uniform prior bound)

lower_start

can be used to set the lower allowable random starting value for the MCMC

upper_start

can be used to set the upper allowable random starting value for the MCMC

par_type

used for bookkeeping during MCMC, set to 1 for normal model parameters, 0 for fixed model options such as exponential_waning, 2 for attack rate terms, phi, or 3 for measurement offset terms, rho

stratification

optional demographic variable or variables used to stratify a parameter

biomarker_group

gives the biomarker group each parameter corresponds to, matching entries in the antibody_data object

See also