Find beta parameters with maximum variance

Given a desired annual mean attack rate and the number of epochs to consider per year, gives the shape1 and shape2 parameters with the desired mean but maximum variance.

find_beta_prior_with_mean(desired_annual_mean, buckets)

Arguments

desired_annual_mean

the desired annual mean attack rate

buckets

the number of time intervals to split each year into

Value

a list with `shape1` and `shape2`

See also

Examples

find_beta_prior_with_mean(0.15, 1)
#> Maximum variance:  0.127499
#> 
#> $shape1
#> [1] 1.17648e-06
#> 
#> $shape2
#> [1] 6.666719e-06
#>