Infection history gibbs proposal

Generates a new infection history matrix and corresponding individual likelihoods, using a gibbs sampler from the infection history prior. See inf_hist_prop_prior_v3, as inputs are very similar.

inf_hist_prop_prior_v2_and_v4(
  theta,
  unique_theta_indices,
  unique_biomarker_groups,
  indiv_group_indices,
  infection_history_mat,
  infection_history_mat_indices,
  likelihoods_pre_proposal,
  sampled_indivs,
  n_times_samp_vec,
  age_mask,
  sample_mask,
  n_alive,
  n_infections,
  n_infected_group,
  prior_lookup,
  proposal_inf_hist_indiv_swap_ratio,
  swap_distance,
  propose_from_prior,
  shape1,
  shape2,
  possible_exposure_times,
  possible_exposure_times_indices,
  exposure_groups,
  unique_exposure_groups,
  sample_times,
  type_data_start,
  biomarker_groups,
  sample_data_start,
  antibody_data_start,
  nrows_per_sample,
  cum_nrows_per_individual_in_data,
  cum_nrows_per_individual_in_repeat_data,
  popn_group_id_vec,
  biomarker_id_indices,
  start_level_indices,
  starting_antibody_levels,
  births,
  antigenic_map_long,
  antigenic_map_short,
  antigenic_distances,
  antibody_data,
  antibody_data_repeats,
  n_measurements_total,
  repeat_indices,
  repeat_data_exist,
  measurement_shifts,
  proposal_iter,
  accepted_iter,
  proposal_swap,
  accepted_swap,
  overall_swap_proposals,
  overall_add_proposals,
  time_sample_probs,
  total_alive,
  data_types,
  obs_weights,
  indiv_possible_exposure_times_indices,
  indiv_poss_exp_times_start,
  indiv_poss_exp_times_end,
  exponential_waning = FALSE,
  timevarying_groups = FALSE,
  variant_specific_pars = FALSE,
  temp = 1,
  solve_likelihood = TRUE
)

Arguments

theta

NumericMatrix, the named model parameters used to solve the model

unique_theta_indices

IntegerVector, the indices of the unique model parameters in the parameter matrix

unique_biomarker_groups

IntegerVector, the biomarker groups represented in the parameter matrix

indiv_group_indices

IntegerVector, for each individual, which unique group (for the kinetics parameters) do they belong to?

infection_history_mat

IntegerMatrix the matrix of 1s and 0s corresponding to individual infection histories

infection_history_mat_indices

IntegerVector, indices linking infection-history entries to the corresponding model inputs

likelihoods_pre_proposal

NumericVector, the current likelihoods for each individual before proposing new infection histories

sampled_indivs

IntegerVector, indices of sampled individuals

n_times_samp_vec

IntegerVector, number of possible exposure times considered for each sampled individual

age_mask

IntegerVector, length of the number of individuals, with indices specifying first time period that an individual can be infected (indexed from 1, such that a value of 1 allows an individual to be infected in any time period)

sample_mask

IntegerVector, length of the number of individuals, with indices specifying last time period that an individual can be infected (ie. last time a sample was taken)

n_alive

IntegerMatrix, number of columns is the number of time periods that an individual could be infected, giving the number of individual alive in each time period. Number of rows is the number of distinct groups.

n_infections

IntegerMatrix, the number of infections in each year (columns) for each group (rows)

n_infected_group

IntegerVector, the total number of infections across all times in each group

prior_lookup

arma::cube, the pre-computed lookup table for the beta prior on infection histories, dimensions are number of infections, time, and group

proposal_inf_hist_indiv_swap_ratio

double, gives the proportion of proposals that will be swap steps (ie. swap contents of two cells in infection_history rather than adding/removing infections)

swap_distance

int, in a swap step, how many time steps either side of the chosen time period to swap with

propose_from_prior

logical, whether to propose infection histories from the prior

shape1

double, shape1 (alpha) parameter for beta prior on infection probability

shape2

double, shape2 (beta) parameter for beta prior on infection probability

possible_exposure_times

NumericVector, the times that individuals could be infected

possible_exposure_times_indices

IntegerVector, indexing vector from 0:(number of exposure times-1)

exposure_groups

IntegerVector, the exposure-group index associated with each possible infection time

unique_exposure_groups

IntegerVector, the exposure groups represented in the model

sample_times

NumericVector, the vector of real times that samples were taken

type_data_start

IntegerVector, the starting index for each individual and biomarker-group block in the antibody data

biomarker_groups

IntegerVector, the biomarker-group index for each individual and sample block

sample_data_start

IntegerVector, the starting index for each individual, biomarker-group, and sample block

antibody_data_start

IntegerVector, the starting index for each individual and biomarker-group block in the antibody data

nrows_per_sample

IntegerVector, the number of antibody rows associated with each sample

cum_nrows_per_individual_in_data

IntegerVector, How many rows in the antibody data correspond to each individual?

cum_nrows_per_individual_in_repeat_data

IntegerVector, For the repeat data (ie. already calculated these antibody levels), how many rows in the antibody data correspond to each individual?

popn_group_id_vec

IntegerVector, vector with 1 entry per individual, giving the group ID of that individual

biomarker_id_indices

IntegerVector, For each antibody measurement, corresponding entry in antigenic map

start_level_indices

IntegerVector, indices linking each individual, biomarker group, and biomarker ID to its starting antibody level

starting_antibody_levels

NumericVector, starting antibody levels for the individual, biomarker group, and biomarker ID combinations

births

NumericVector, birth times for each individual

antigenic_map_long

arma::mat, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1, see create_cross_reactivity_vector

antigenic_map_short

arma::mat, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2, see create_cross_reactivity_vector

antigenic_distances

NumericVector matching the dimensions of antigenic_map_long and antigenic_map_short, but with the raw antigenic distances between strains

antibody_data

NumericVector, data for all individuals for the first instance of each calculated antibody level

antibody_data_repeats

NumericVector, the repeat antibody data for all individuals (ie. do not solve the same antibody level twice)

n_measurements_total

integer, total number of antibody measurements represented in the flattened data

repeat_indices

IntegerVector, which index in the main data vector does each entry in repeat_data correspond to ie. which calculated antibody level in predicted_antibody_levels should be used for each observation?

repeat_data_exist

logical, whether repeat antibody data are present

measurement_shifts

NumericVector, if length matches the length of data, adds these as measurement shifts to the antibody levels. If lengths do not match, is not used.

proposal_iter

IntegerVector, vector with entry for each individual, storing the number of infection history add/remove proposals for each individual.

accepted_iter

IntegerVector, vector with entry for each individual, storing the number of accepted infection history add/remove proposals for each individual.

proposal_swap

IntegerVector, vector with entry for each individual, storing the number of proposed infection history swaps

accepted_swap

IntegerVector, vector with entry for each individual, storing the number of accepted infection history swaps

overall_swap_proposals

IntegerMatrix, counts of proposed swaps by group and time

overall_add_proposals

IntegerMatrix, counts of proposed additions or removals by group and time

time_sample_probs

NumericVector, sampling probabilities for possible exposure times

total_alive

IntegerVector, giving the total number of potential infection events for each group. This only applies to prior version 4. If set to a vector of values -1, then this is ignored.

data_types

IntegerVector, observation-model type for each biomarker group

obs_weights

NumericVector, observation weights used in the likelihood

indiv_possible_exposure_times_indices

IntegerVector, indices of possible exposure times available to each individual

indiv_poss_exp_times_start

IntegerVector, starting index for each individual's possible exposure times

indiv_poss_exp_times_end

IntegerVector, ending index for each individual's possible exposure times

exponential_waning

logical, whether to use the exponential rather than linear waning and cross-reactivity form

timevarying_groups

logical, whether demographic groups can vary over time

variant_specific_pars

logical, whether parameter blocks are selected by exposure group

temp

double, temperature for parallel tempering MCMC

solve_likelihood

logical, if FALSE does not solve the likelihood when calculating the acceptance probability

Value

an R list with 6 entries: 1) the vector replacing likelihoods_pre_proposal, corresponding to the new likelihoods per individual; 2) the matrix of 1s and 0s corresponding to the new infection histories for all individuals; 3-6) the updated entries for proposal_iter, accepted_iter, proposal_swap and accepted_swap.

See also