Prior on measurement shifts

Assumes measurement shifts are drawn from a normal distribution (random effects) with given standard deviation and mean. Code is commented out to assume log normally distributied

prob_shifts(rhos, pars)

Arguments

rhos

vector of measurement shifts

pars

vector of parameters, including rho_mean and rho_sd for the normal distribution

Value

a single log prior probability

See also