Example parameter table

Example of the parameter table used through serosolver. This data frame is used to control everything related to the model parameters, including values, names, fixed/estimates, and uniform prior bounds.

data(example_par_tab)

Format

A data frame with 21 rows and 9 variables:

names

string names of the model parameters

values

numeric values of the parameters

fixed

binary values indicating if the parameter should be fixed (1) or estimated (0) during the MCMC procedure

steps

value between 0 and 1, giving the initial step size in the MCMC proposals. Note that these are adapted automatically

lower_bound

lower numeric bound for the parameter during fitting (lower uniform prior bound)

upper_bound

upper numeric bound for the parameter during fitting (upper uniform prior bound)

lower_start

can be used to set the lower allowable random starting value for the MCMC

upper_start

can be used to set the upper allowable random starting value for the MCMC

type

used for admin during MCMC, set to 1 for normal model parameters, 0 for model options (eg. see titre_dependent), or 2 for attack rate terms, phi

See also